requirements.txt
django==4.2.3
djangorestframework==3.14.0
drf-spectacular==0.26.3
djoser=2.2.0
